# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentification réussie"
          },
          "401": {
            "description": "Authentification échouée"
          }
        },
        "summary": "Authentification d'un administrateur",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profil récupéré avec succès"
          },
          "401": {
            "description": "Non autorisé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer le profil de l'utilisateur connecté",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/admin/database/stats": {
      "get": {
        "operationId": "DatabaseController_getDatabaseStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Statistiques récupérées avec succès"
          },
          "401": {
            "description": "Non autorisé"
          },
          "403": {
            "description": "Accès interdit"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer les statistiques de la base de données (super admin uniquement)",
        "tags": [
          "database"
        ]
      }
    },
    "/api/admin/database/export": {
      "post": {
        "operationId": "DatabaseController_exportDatabase",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Base de données exportée avec succès"
          },
          "401": {
            "description": "Non autorisé"
          },
          "403": {
            "description": "Accès interdit"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Exporter la base de données (super admin uniquement)",
        "tags": [
          "database"
        ]
      }
    },
    "/api/admin/database/import": {
      "post": {
        "operationId": "DatabaseController_importDatabase",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Base de données importée avec succès"
          },
          "400": {
            "description": "Fichier invalide"
          },
          "401": {
            "description": "Non autorisé"
          },
          "403": {
            "description": "Accès interdit"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Importer la base de données (super admin uniquement)",
        "tags": [
          "database"
        ]
      }
    },
    "/api/admin/database/reset": {
      "post": {
        "operationId": "DatabaseController_resetDatabase",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Base de données réinitialisée avec succès"
          },
          "401": {
            "description": "Non autorisé"
          },
          "403": {
            "description": "Accès interdit"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Réinitialiser la base de données (super admin uniquement)",
        "tags": [
          "database"
        ]
      }
    },
    "/api/admin/dashboard/stats": {
      "get": {
        "operationId": "DashboardController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Statistiques récupérées avec succès"
          },
          "401": {
            "description": "Non autorisé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer les statistiques pour le tableau de bord",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/api/admin/dashboard/yearly-stats": {
      "get": {
        "operationId": "DashboardController_getYearlyStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Statistiques annuelles récupérées avec succès"
          },
          "401": {
            "description": "Non autorisé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer les statistiques annuelles",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/api/admin/dashboard/inscription-trends": {
      "get": {
        "operationId": "DashboardController_getInscriptionTrends",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Tendances récupérées avec succès"
          },
          "401": {
            "description": "Non autorisé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer les tendances d'inscription",
        "tags": [
          "dashboard"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Aloha Secourisme",
    "description": "\n      API REST complète pour la gestion de l'application Aloha Secourisme.\n      \n      Cette API permet de gérer les formations, les inscriptions, les documents, \n      la galerie d'images et l'administration du système.\n      \n      ## Authentification\n      \n      La plupart des endpoints nécessitent une authentification par token JWT.\n      Pour vous authentifier, utilisez l'endpoint `/api/auth/login` pour obtenir un token,\n      puis incluez ce token dans l'en-tête d'autorisation de vos requêtes.\n      \n      ## Rôles\n      \n      - **Accès public** : Accessible sans authentification\n      - **Admin** : Nécessite d'être connecté avec un compte administrateur\n      - **Super Admin** : Nécessite d'être connecté avec un compte super administrateur\n      \n      ## Modules principaux\n      \n      - **Formations** : Gestion des sessions de formation\n      - **Inscriptions** : Gestion des inscriptions aux formations\n      - **Documents** : Gestion des documents téléchargeables\n      - **Galerie** : Gestion des images de la galerie\n      - **News** : Gestion des actualités\n      - **Admin/Users** : Gestion des utilisateurs administrateurs\n      - **Dashboard** : Statistiques et tableaux de bord\n      - **Settings** : Paramètres du système\n      - **Database** : Export/import de la base de données\n    ",
    "version": "1.0.0",
    "contact": {
...
    },
...
    }
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Entrez votre token JWT ici"
      }
    },
    "schemas": {
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@example.com",
            "description": "Adresse email de l'utilisateur"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "Mot de passe de l'utilisateur"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    }
  }
}